% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-jitter2D.R
\name{geom_jitter2D}
\alias{geom_jitter2D}
\title{Jittered 2D points}
\usage{
geom_jitter2D(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "jitter2D",
  ...,
  weight = NULL,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{weight}{Amount of vertical and horizontal jitter. The jitter
  is added in both positive and negative directions, so the total spread
  is twice the value specified here.

  If omitted, defaults to 40% of the resolution of the data: this means the
  jitter values will occupy 80% of the implied bins. Categorical data
  is aligned on the integers, so a width or height of 0.5 will spread the
  data so it's not possible to see the distinction between the categories.}
}
\description{
The jitter geom is a convenient shortcut for
`geom_point(position = "jitter2D")`. It adds a small amount of random
variation to the location of each point using bivarite normal distribution, and is a useful way of handling
overplotting caused by discreteness in smaller datasets.
}
\seealso{
[geom_point()] for regular, unjittered points,
 [geom_boxplot()] for another way of looking at the conditional
    distribution of a variable
}
