% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-jitter-gauss.R
\docType{data}
\name{position_jitter_gauss}
\alias{position_jitter_gauss}
\alias{PositionJittergauss}
\title{2D Jitter points to avoid overplotting}
\usage{
position_jitter_gauss(weight = NULL, seed = NA)
}
\arguments{
\item{weight, }{Amount of vertical and horizontal jitter. The jitter
  is added in both positive and negative directions, so the total spread
  is twice the value specified here.

  If omitted, defaults to 40% of the resolution of the data: this means the
  jitter values will occupy 80% of the implied bins. Categorical data
  is aligned on the integers, so a width or height of 0.5 will spread the
  data so it's not possible to see the distinction between the categories.}

\item{seed}{A random seed to make the jitter reproducible.
Useful if you need to apply the same jitter twice, e.g., for a point and
a corresponding label.
The random seed is reset after jittering.
If `NA` (the default value), the seed is initialised with a random value;
this makes sure that two subsequent calls start with a different seed.
Use `NULL` to use the current random seed and also avoid resetting
(the behaviour of \pkg{ggplot} 2.2.1 and earlier).}
}
\description{
2D Jitter points to avoid overplotting
}
\concept{position adjustments}
\keyword{datasets}
