% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-jitter-quasiloc.R
\docType{data}
\name{position_jitter_quasiloc}
\alias{position_jitter_quasiloc}
\alias{PositionJitterquasiloc}
\title{2D Jitter points to avoid overplotting based on Sobolev local sequence}
\usage{
position_jitter_quasiloc(weight = NULL, seed = NA)
}
\arguments{
\item{weight, }{spread factor.

  If omitted, just include the spread in the data.}

\item{seed}{A random seed to make the jitter reproducible.
Useful if you need to apply the same jitter twice, e.g., for a point and
a corresponding label.
The random seed is reset after jittering.
If `NA` (the default value), the seed is initialised with a random value;
this makes sure that two subsequent calls start with a different seed.
Use `NULL` to use the current random seed and also avoid resetting
(the behaviour of \pkg{ggplot} 2.2.1 and earlier).}

\item{loc}{logical, if its TRUE it uses the Sobol sequence to generate points locally, and it is false generate a complete sobol sequence for all points.}
}
\description{
2D Jitter points to avoid overplotting based on Sobolev local sequence
}
\examples{
ggplot2::mpg |> ggplot2::ggplot(ggplot2::aes(cty, hwy)) +
 geom_jitter_quasiloc() + ggplot2::theme(aspect.ratio = 1)

}
\seealso{
Other position adjustments: 
\code{\link{position_jitter_gauss}()},
\code{\link{position_jitter_quasi}()},
\code{\link{position_jitter_quasitog}()}
}
\concept{position adjustments}
\keyword{datasets}
