% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-jitter-quasitog.R
\name{position_jitter_quasitog}
\alias{position_jitter_quasitog}
\title{2D Jitter points to avoid overplotting based on Sobol sequence (toggle local/global)}
\usage{
position_jitter_quasitog(weight = NULL, seed = NA, loc = FALSE)
}
\arguments{
\item{weight}{Spread factor.
If omitted, it is automatically set from the data.}

\item{seed}{Random seed for reproducibility.}

\item{loc}{Logical. If TRUE, generates points locally per (x,y) duplicate.
If FALSE, generates a global Sobol sequence for all points.}
}
\description{
2D Jitter points to avoid overplotting based on Sobol sequence (toggle local/global)
}
\seealso{
Other position adjustments: 
\code{\link{position_jitter_gauss}()}
}
\concept{position adjustments}
