% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_jitter2D.R
\name{geom_jitter2D}
\alias{geom_jitter2D}
\title{Jittered 2D points}
\usage{
geom_jitter2D(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "jitter",
  ...,
  width = NULL,
  height = NULL,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\description{
The jitter geom is a convenient shortcut for
`geom_point(position = "jitter2D")`. It adds a small amount of random
variation to the location of each point using bivarite normal distribution, and is a useful way of handling
overplotting caused by discreteness in smaller datasets.
}
\examples{
p <- ggplot(mpg, aes(cty, hwy))
p + geom_point()
p + geom_jitter2D()

# Add aesthetic mappings
p + geom_jitter2D(aes(colour = class))
}
\seealso{
[geom_point()] for regular, unjittered points,
 [geom_boxplot()] for another way of looking at the conditional
    distribution of a variable
}
