% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-jitter-gauss.R
\docType{data}
\name{position_jitter_gauss}
\alias{position_jitter_gauss}
\alias{PositionJittergauss}
\title{2D Jitter points to avoid overplotting based on Gaussian bivariate distribution}
\usage{
position_jitter_gauss(weight = NULL, seed = NA)
}
\arguments{
\item{weight, }{spread factor.

  If omitted, just include the spread in the data.}

\item{seed}{A random seed to make the jitter reproducible.
Useful if you need to apply the same jitter twice, e.g., for a point and
a corresponding label.
The random seed is reset after jittering.
If `NA` (the default value), the seed is initialised with a random value;
this makes sure that two subsequent calls start with a different seed.
Use `NULL` to use the current random seed and also avoid resetting
(the behaviour of \pkg{ggplot} 2.2.1 and earlier).}
}
\description{
2D Jitter points to avoid overplotting based on Gaussian bivariate distribution
}
\examples{
ggplot2::mpg |> ggplot2::ggplot(ggplot2::aes(cty, hwy)) +
 geom_jitter_gauss() + ggplot2::theme(aspect.ratio = 1)
}
\seealso{
Other position adjustments: 
\code{\link{position_jitter_quasi}()},
\code{\link{position_jitter_quasiloc}()},
\code{\link{position_jitter_quasitog}()}
}
\concept{position adjustments}
\keyword{datasets}
